fix: rename Postgres to Postgre where it makes sense
