+++ src/Conways.GameOfLife.Infrastructure.PostgreSQL/Configurations/GenerationEntityTypeConfiguration.cs
@@ -2,7 +2,7 @@ using Conways.GameOfLife.Domain;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace Conways.GameOfLife.Infrastructure.PostgresSQL.Configurations;
+namespace Conways.GameOfLife.Infrastructure.PostgreSQL.Configurations;
 
 public class GenerationEntityTypeConfiguration : IEntityTypeConfiguration<Generation>
 {
