+++ src/Conways.GameOfLife.Infrastructure.PostgreSQL/Extensions/ServiceCollectionExtensions.cs
@@ -1,9 +1,9 @@
-using Conways.GameOfLife.Infrastructure.PostgresSQL.Interceptors;
+using Conways.GameOfLife.Infrastructure.PostgreSQL.Interceptors;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Conways.GameOfLife.Infrastructure.PostgresSQL.Extensions;
+namespace Conways.GameOfLife.Infrastructure.PostgreSQL.Extensions;
 
 public static class ServiceCollectionExtensions
 {
