+++ tests/Conways.GameOfLife.IntegrationTests/GlobalUsings.cs
@@ -6,7 +6,6 @@ global using Conways.GameOfLife.API.Features.UploadBoard;
 global using Conways.GameOfLife.API.Models;
 global using Conways.GameOfLife.Domain.Exceptions;
 global using Conways.GameOfLife.Domain;
-global using Conways.GameOfLife.Infrastructure.PostgresSQL.Interceptors;
 global using Conways.GameOfLife.Infrastructure.PostgresSQL;
 global using Conways.GameOfLife.IntegrationTests.Factories;
 global using FluentAssertions;
