+++ tests/Conways.GameOfLife.IntegrationTests/Postgre/BoardDbContextTests.cs
@@ -1,4 +1,6 @@
-namespace Conways.GameOfLife.IntegrationTests.Postgres;
+using Conways.GameOfLife.Infrastructure.PostgreSQL;
+
+namespace Conways.GameOfLife.IntegrationTests.Postgre;
 
 public class BoardDbContextTests : IClassFixture<ConwaysGameOfLifeWebApplicationFactory>
 {
